BEC_API/requirements.txt
fastapi
uvicorn[standard]
PyJWT
sqlalchemy
psycopg2-binary #PostreSQL
python-jose[cryptography] #generar, firmar y dedodificar los tokens
passlib[bcrypt] #encriptar y verificar contraseñas
bcrypt==3.2.2 # Fix para passlib error de compatibilidad
python-multipart #leer formulafio de login de OAuth2
pydantic[email]